scoring-engine/requirements.txt
# ARC Scoring Engine - Dependencies

# Core
pandas>=1.5.0
numpy>=1.24.0
scipy>=1.10.0
pyyaml>=6.0

# Database
psycopg2-binary>=2.9.0

# Data Sources
yfinance>=0.2.0
polygon-api-client>=1.0.0
requests>=2.28.0

# Data Processing
pyarrow>=12.0.0
fastparquet>=2023.0.0

# Logging
python-json-logger>=2.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Type Hints
typing-extensions>=4.0.0
